Adding register, login, logout

+++ csgostash/base/views.py
@@ -1,13 +1,72 @@
 from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Weapon
-from django.http import HttpResponseNotFound
+from django.http import HttpResponseNotFound, HttpResponse
 from .forms import WeaponForm
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate,login,logout
+from django.contrib.auth.forms import UserCreationForm
+
 
 # Create your views here.
 
 
+def loginPage(request):
+
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == "POST":
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist!")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        
+        else:
+            messages.error(request, "Username or Password does not exist!")
+
+
+    context = {'page': page}
+    return render(request, 'login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occurred during registration")
+    return render(request, 'login_register.html', {'form': form})
+
+
 def home(request):
-    return render(request, 'home.html')
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+    weapons = Weapon.objects.filter(skin_name__icontains = q)
+    context = {'weapons': weapons}
+    return render(request, 'existing_skins.html', context)
 
 def rifles(request):
     return render(request, 'rifles.html')
@@ -51,7 +110,7 @@ def current_rifle(request,weapon, pk):
         context = {'weapon': weapon_from_database.get()}
         return render(request, 'current_rifle.html', context)
 
-
+@login_required(login_url='/login')
 def sell_item(request):
     form = WeaponForm()
     if request.method == 'POST':
@@ -64,9 +123,13 @@ def sell_item(request):
     return render(request, 'sell.html', context)
 
 
+@login_required(login_url='/login')
 def update_item(request, pk):
     weapon = Weapon.objects.get(id=pk)
     form = WeaponForm(instance=weapon)
+
+    if request.user != weapon.seller:
+        return HttpResponse('Cannot edit this Ad!')
 
     if request.method == 'POST':
         form = WeaponForm(request.POST, instance=weapon)
@@ -78,8 +141,13 @@ def update_item(request, pk):
     return render(request, 'sell.html', context)
 
 
+@login_required(login_url='/login')
 def delete_item(request, pk):
     item = Weapon.objects.get(id=pk)
+
+    if request.user != item.seller:
+        return HttpResponse('Cannot edit this Ad!')
+    
     if request.method == 'POST':
         item.delete()
         return redirect('home')
